MM_POC_v2_video_gen/nanobana2-engineer-package (4)/pyproject.toml
[project]
name = "nanobana2"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "pillow",
    "pydantic",
    "pinterest-dl",
    "python-dotenv",
    "pyyaml",
    "httpx",
]
